refactor into commands and classes, add docs
This commit refactors how commands are defined to leverage yargs cli
documentation features. A clean split is defined between classes where
each commands logic is defined and commands where user input is parsed
and passed to the appropriate class.

classes are exported for programmatic usage, when the main module is
executed directly, it acts as CLI