fix: incorrect version number displayed on alias when updating with a cache layer infront
With a cache layer, such as Fastly, the version number returned on the GET is not always what we just set, so it can cause some confusion. Now we're using the direct input rather than response for a more accurate version number displayed in the Eik response on alias update.