feat: add alias command as an eventual replacement for the 3 existing alias commands
This is a non breaking change but does add deprecation messages for pkg-alias, map-alias and npm-alias, all of which
are replaced by this new alias command.

The alias command detects type from eik.json which was not possible in earlier versions of the cli which was the reason for the existence of 3 separate commands instead of 1.
If the command is run in a directory that does not contain an eik.json file, type can be provided via the command line argument --type. If neither are present, an error is thrown.